Dependency Kurtosis Client update. Test renamed to WaitForEdnpointAvailabilityTest
