+++ golang/go.mod
@@ -3,7 +3,7 @@ module github.com/kurtosis-tech/kurtosis-libs/golang
 go 1.13
 
 require (
-	github.com/kurtosis-tech/kurtosis-client/golang v0.0.0-20210625134328-8c47a112180e
+	github.com/kurtosis-tech/kurtosis-client/golang v0.0.0-20210625204610-dbc8ce488696
 	github.com/kurtosis-tech/minimal-grpc-server v0.0.0-20210504182615-82226e94877b
 	github.com/palantir/stacktrace v0.0.0-20161112013806-78658fd2d177
 	github.com/sirupsen/logrus v1.8.1
